Refactoring
Moved gallery components into a sub-folder
