+++ src/components/gallery/galleryItem.js
