+++ src/components/gallery/galleryItemSelected.js
