+++ src/components/gallery/galleryItemSelected.module.css
