+++ src/pages/media.js
@@ -1,7 +1,7 @@
 import React from "react"
 import { graphql } from "gatsby"
 import Layout from "../components/layout"
-import Gallery from "../components/gallery"
+import Gallery from "../components/gallery/gallery"
 import SEO from "../components/seo"
 
 /**
